requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
pydantic==1.10.13 